lxbuildenv: avoid specific _sim version, use the same lxbuildenv.py for build and simulation
